+++ src/mpd_client/mpd_client_jukebox.c
@@ -242,7 +242,7 @@ bool mpd_client_jukebox_add_to_queue(struct t_mympd_state *mympd_state, long add
     if (manual == false) {
         MYMPD_LOG_DEBUG("Jukebox queue length: %u", mympd_state->jukebox_queue.length);
     }
-    if ((manual == false && add_songs > (long)mympd_state->jukebox_queue.length) ||
+    if ((manual == false && add_songs > mympd_state->jukebox_queue.length) ||
         (manual == true))
     {
         bool rc = mpd_client_jukebox_fill_jukebox_queue(mympd_state, add_songs, jukebox_mode, playlist, manual);
